Solved a critical issue for drawing.
Solved a critical issue where writing to the displayAddress caused me to overwrite the buffer values. This is because displayAddress had half the memory allocated for the pixels than required, which is why I added an extra variable to set aside extra space for the display to use. I took advantage of the contiguous memory aspect of it.